misc/closed-web-net-2/attachments/pyproject.toml

[project]
name = "attachments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "pyown>=0.3.6",
    "pyzbar>=0.1.9",
    "Pillow>=11.3.0",
]
